fix float coords

+++ backend/internal/projects/handlers.go
@@ -36,8 +36,8 @@ type createProjectResponse struct {
 }
 
 type Coordinate struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type createBuildingInput struct {
